LPS-54744 Move everything into the correct place
